backend/requirements.txt
# Stockman - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Data Sources
yfinance==0.2.36
finnhub-python==2.4.19
alpha-vantage==2.3.1
feedparser==6.0.10
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3

# AI/ML
anthropic==0.40.0
openai==1.12.0  # For Whisper API

# Voice
elevenlabs==1.0.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
pytz==2024.1
